Uploading the form.
Almost done with main Task form that will be used to operate the tasks and collaborating with App.js to functionalize the required tasks. Facing code errors with the task list files that will be developed to show the user task files. Using Youtube links to get the results,

https://www.youtube.com/watch?v=yU7U5fpExxs
https://www.youtube.com/watch?v=MINQoCSBmds
https://www.youtube.com/watch?v=2aIjW_w_-Rg

Also working with storing the data that is JSON.stringify and JSON.parce for the search engines 
Using get.element(), set.element(), clear() functions to reach the destination.

I have a target to bonus task , at least implementing the dark mood.